Auth/requirements.txt
annotated-types==0.6.0
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
email-validator==2.1.0.post1
Flask==3.0.2
Flask-PyMongo==2.3.0
Flask-WTF==1.2.1
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
passlib==1.7.4
pydantic==2.6.2
pydantic_core==2.16.3
PyJWT==2.8.0
pykerberos==1.2.4
pymongo==4.6.2
python-dotenv==1.0.1
python-snappy==0.6.1
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
zipp==3.17.0
